SO-4100: extract inactivation/reactivation test cases to another class
